refactor(web): extract shared billing-checkout mutation hook
